优化AI功能和用户体验
主要改进:
- AI功能修复: 恢复立即初始化AI配置，移除延迟
- 新手引导优化: 修复引导完成后转圈问题，优化加载状态
- AI回复优化: 使用Markdown渲染，回复更简洁结构化
- 食物分析增强: 图片和文本导入均并发调用5次取平均值
- UI修复: 修复底部选项卡描述文字显示问题
- 代码清理: 移除未使用的导入和参数

+++ CalorieAI/app/src/main/java/com/calorieai/app/ui/screens/camera/PhotoAnalysisViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.calorieai.app.data.model.FoodRecord
 import com.calorieai.app.data.model.MealType
+import com.calorieai.app.data.model.FoodAnalysisResult
 import com.calorieai.app.data.repository.FoodRecordRepository
-import com.calorieai.app.service.ai.FoodAnalysisResult
 import com.calorieai.app.service.ai.FoodImageAnalysisService
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -97,7 +97,7 @@ class PhotoAnalysisViewModel @Inject constructor(
             val record = FoodRecord(
                 foodName = result.foodName,
                 userInput = "拍照识别：${result.foodName} - ${result.description}",
-                totalCalories = result.calories,
+                totalCalories = result.calories.toInt(),
                 protein = result.protein,
                 carbs = result.carbs,
                 fat = result.fat,
